requirements.txt
plexapi==4.16.1
yt-dlp==2025.01.12
PyYAML==6.0.2